Refactor shortcodes and filters in separate files
